the matrix can draw the dots with varying lightness by using a float 0.0-1.1

+++ selfdrive/ui/qt/body.h
@@ -15,7 +15,7 @@ public:
   int dotSize;
 
 protected:
-  void paintMatrix(int a[4][4]);
+  void paintMatrix(float a[4][4]);
   virtual void paintEvent(QPaintEvent *event);
 };
